+++ GAAL-Workspace/Game-As-A-Lab-Step-04-score-and-events/src/rollball/model/GameObject.java
@@ -30,7 +30,7 @@ public class GameObject  {
 		this.vel = new V2d(-vel.x, vel.y);
 	}
 	
-	public void updateState(int dt){
+	public void updateState(long dt){
 		this.pos = this.pos.sum(vel.mul(0.001*dt));
 	}
