+++ GAAL-Workspace/Game-As-A-Lab-Step-07-mosquito-AI/src/rollball/model/GameObject.java
@@ -64,7 +64,7 @@ public class GameObject {
 		input.update(this, c);
 	}
 
-	public void updatePhysics(int dt, World w){
+	public void updatePhysics(long dt, World w){
 		phys.update(dt, this, w);
 	}
