+++ GAAL-Workspace/Game-As-A-Lab-Step-11-multiplayer-remote-users/src/rollball/physics/PhysicsComponent.java
@@ -6,7 +6,7 @@ import rollball.model.*;
 
 public abstract class PhysicsComponent {
 
-	public void update(int dt, GameObject obj, World w){
+	public void update(long dt, GameObject obj, World w){
 		P2d pos = obj.getCurrentPos();
 		V2d vel = obj.getCurrentVel();
 		obj.setPos(pos.sum(vel.mul(0.001*dt)));
