+++ GAAL-Workspace/Game-As-A-Lab-Step-12-multiplayer-remote-users-with-view/src/rollball/model/World.java
@@ -45,7 +45,7 @@ public class World {
 		picks.remove(obj);
 	}
 	
-	public void updateState(int dt){
+	public void updateState(long dt){
 		for (GameObject ball: balls) {
 			ball.updatePhysics(dt, this);
 		}
