+++ GAAL-Workspace/Game-As-A-Lab-Step-12-multiplayer-remote-users-with-view/src/rollball/physics/PickUpObjPhysicsComponent.java
@@ -5,7 +5,7 @@ import rollball.model.World;
 
 public class PickUpObjPhysicsComponent extends PhysicsComponent {
 	
-	public void update(int dt, GameObject obj, World w) {
+	public void update(long dt, GameObject obj, World w) {
 	}
 
 }
